Reduce code duplicates in topology plugins
Reduce code duplicates in topology plugins by introducing common
abstract parent which performs common operations to get netconf base
and netconf nmda services.

JIRA: LIGHTY-87
