+++ lighty-modules/lighty-netconf-sb/src/main/java/io/lighty/modules/southbound/netconf/impl/NetconfClusteredTopologyPlugin.java
@@ -7,16 +7,9 @@
  */
 package io.lighty.modules.southbound.netconf.impl;
 
-import io.lighty.core.controller.api.AbstractLightyModule;
 import io.lighty.core.controller.api.LightyServices;
-import io.lighty.modules.southbound.netconf.impl.util.NetconfUtils;
-import java.util.Optional;
 import java.util.concurrent.ExecutorService;
 import org.opendaylight.aaa.encrypt.AAAEncryptionService;
-import org.opendaylight.mdsal.dom.api.DOMMountPoint;
-import org.opendaylight.mdsal.dom.api.DOMMountPointService;
-import org.opendaylight.mdsal.dom.api.DOMRpcService;
-import org.opendaylight.mdsal.dom.api.DOMSchemaService;
 import org.opendaylight.netconf.client.NetconfClientDispatcher;
 import org.opendaylight.netconf.sal.connect.api.SchemaResourceManager;
 import org.opendaylight.netconf.sal.connect.impl.DefaultSchemaResourceManager;
@@ -25,18 +18,14 @@ import org.opendaylight.netconf.sal.connect.netconf.schema.mapping.DefaultBaseNe
 import org.opendaylight.netconf.topology.singleton.impl.NetconfTopologyManager;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.netconf.topology.singleton.config.rev170419.Config;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.netconf.topology.singleton.config.rev170419.ConfigBuilder;
-import org.opendaylight.yang.gen.v1.urn.tbd.params.xml.ns.yang.network.topology.rev131021.NodeId;
 import org.opendaylight.yangtools.yang.common.Uint16;
-import org.opendaylight.yangtools.yang.data.api.YangInstanceIdentifier;
 import org.opendaylight.yangtools.yang.parser.api.YangParserException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class NetconfClusteredTopologyPlugin extends AbstractLightyModule implements NetconfSBPlugin {
-
+public final class NetconfClusteredTopologyPlugin extends AbstractTopologyPlugin {
     private static final Logger LOG = LoggerFactory.getLogger(NetconfClusteredTopologyPlugin.class);
 
-    private final DOMMountPointService domMountPointService;
     private final LightyServices lightyServices;
     private final String topologyId;
     private final NetconfClientDispatcher clientDispatcher;
@@ -46,8 +35,7 @@ public class NetconfClusteredTopologyPlugin extends AbstractLightyModule impleme
     public NetconfClusteredTopologyPlugin(final LightyServices lightyServices, final String topologyId,
             final NetconfClientDispatcher clientDispatcher, final Integer writeTxIdleTimeout,
             final ExecutorService executorService, final AAAEncryptionService encryptionService) {
-        super(executorService);
-        this.domMountPointService = lightyServices.getDOMMountPointService();
+        super(executorService, lightyServices.getDOMMountPointService());
         this.lightyServices = lightyServices;
         this.topologyId = topologyId;
         this.clientDispatcher = clientDispatcher;
@@ -89,40 +77,5 @@ public class NetconfClusteredTopologyPlugin extends AbstractLightyModule impleme
     @Override
     public boolean isClustered() {
         return true;
-    }
-
-    @Override
-    public Optional<NetconfBaseService> getNetconfBaseService(final NodeId nodeId) {
-        final Optional<DOMMountPoint> domMountPointOptional = getNetconfDOMMountPoint(nodeId);
-        if (domMountPointOptional.isPresent()) {
-            final DOMMountPoint domMountPoint = domMountPointOptional.get();
-            Optional<DOMSchemaService> service = domMountPoint.getService(DOMSchemaService.class);
-            final Optional<DOMRpcService> domRpcServiceOptional = domMountPoint.getService(DOMRpcService.class);
-            if (domRpcServiceOptional.isPresent()) {
-                return Optional.of(new NetconfBaseServiceImpl(nodeId, domRpcServiceOptional.get(),
-                        service.orElseThrow().getGlobalContext()));
-            }
-        }
-        return Optional.empty();
-    }
-
-    @Override
-    public Optional<NetconfNmdaBaseService> getNetconfNmdaBaseService(final NodeId nodeId) {
-        final Optional<DOMMountPoint> domMountPointOptional = getNetconfDOMMountPoint(nodeId);
-        if (domMountPointOptional.isPresent()) {
-            final DOMMountPoint domMountPoint = domMountPointOptional.get();
-            Optional<DOMSchemaService> service = domMountPoint.getService(DOMSchemaService.class);
-            final Optional<DOMRpcService> domRpcServiceOptional = domMountPoint.getService(DOMRpcService.class);
-            if (domRpcServiceOptional.isPresent()) {
-                return Optional.of(new NetconfNmdaBaseServiceImpl(nodeId, domRpcServiceOptional.get(),
-                        service.orElseThrow().getGlobalContext()));
-            }
-        }
-        return Optional.empty();
-    }
-
-    private Optional<DOMMountPoint> getNetconfDOMMountPoint(final NodeId nodeId) {
-        final YangInstanceIdentifier yangInstanceIdentifier = NetconfUtils.createNetConfNodeMountPointYII(nodeId);
-        return this.domMountPointService.getMountPoint(yangInstanceIdentifier);
     }
 }
