+++ lighty-modules/lighty-netconf-sb/src/main/java/io/lighty/modules/southbound/netconf/impl/NetconfTopologyPlugin.java
@@ -7,32 +7,21 @@
  */
 package io.lighty.modules.southbound.netconf.impl;
 
-import io.lighty.core.controller.api.AbstractLightyModule;
 import io.lighty.core.controller.api.LightyServices;
-import io.lighty.modules.southbound.netconf.impl.util.NetconfUtils;
-import java.util.Optional;
 import java.util.concurrent.ExecutorService;
 import org.opendaylight.aaa.encrypt.AAAEncryptionService;
-import org.opendaylight.mdsal.dom.api.DOMMountPoint;
-import org.opendaylight.mdsal.dom.api.DOMMountPointService;
-import org.opendaylight.mdsal.dom.api.DOMRpcService;
-import org.opendaylight.mdsal.dom.api.DOMSchemaService;
 import org.opendaylight.netconf.client.NetconfClientDispatcher;
 import org.opendaylight.netconf.sal.connect.api.SchemaResourceManager;
 import org.opendaylight.netconf.sal.connect.impl.DefaultSchemaResourceManager;
 import org.opendaylight.netconf.sal.connect.netconf.schema.mapping.DefaultBaseNetconfSchemas;
 import org.opendaylight.netconf.topology.impl.NetconfTopologyImpl;
-import org.opendaylight.yang.gen.v1.urn.tbd.params.xml.ns.yang.network.topology.rev131021.NodeId;
-import org.opendaylight.yangtools.yang.data.api.YangInstanceIdentifier;
 import org.opendaylight.yangtools.yang.parser.api.YangParserException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class NetconfTopologyPlugin extends AbstractLightyModule implements NetconfSBPlugin {
-
+public final class NetconfTopologyPlugin extends AbstractTopologyPlugin {
     private static final Logger LOG = LoggerFactory.getLogger(NetconfTopologyPlugin.class);
 
-    private final DOMMountPointService domMountPointService;
     private final String topologyId;
     private final NetconfClientDispatcher clientDispatcher;
     private NetconfTopologyImpl netconfTopologyImpl;
@@ -42,8 +31,7 @@ public class NetconfTopologyPlugin extends AbstractLightyModule implements Netco
     NetconfTopologyPlugin(final LightyServices lightyServices, final String topologyId,
             final NetconfClientDispatcher clientDispatcher, final ExecutorService executorService,
             final AAAEncryptionService encryptionService) {
-        super(executorService);
-        this.domMountPointService = lightyServices.getDOMMountPointService();
+        super(executorService, lightyServices.getDOMMountPointService());
         this.lightyServices = lightyServices;
         this.topologyId = topologyId;
         this.clientDispatcher = clientDispatcher;
@@ -82,41 +70,5 @@ public class NetconfTopologyPlugin extends AbstractLightyModule implements Netco
     @Override
     public boolean isClustered() {
         return false;
-    }
-
-    @Override
-    public Optional<NetconfBaseService> getNetconfBaseService(final NodeId nodeId) {
-
-        final Optional<DOMMountPoint> domMountPointOptional = getNetconfDOMMountPoint(nodeId);
-        if (domMountPointOptional.isPresent()) {
-            final DOMMountPoint domMountPoint = domMountPointOptional.get();
-            Optional<DOMSchemaService> service = domMountPoint.getService(DOMSchemaService.class);
-            final Optional<DOMRpcService> domRpcServiceOptional = domMountPoint.getService(DOMRpcService.class);
-            if (domRpcServiceOptional.isPresent()) {
-                return Optional.of(new NetconfBaseServiceImpl(nodeId, domRpcServiceOptional.get(),
-                        service.orElseThrow().getGlobalContext()));
-            }
-        }
-        return Optional.empty();
-    }
-
-    @Override
-    public Optional<NetconfNmdaBaseService> getNetconfNmdaBaseService(NodeId nodeId) {
-        final Optional<DOMMountPoint> domMountPointOptional = getNetconfDOMMountPoint(nodeId);
-        if (domMountPointOptional.isPresent()) {
-            final DOMMountPoint domMountPoint = domMountPointOptional.get();
-            Optional<DOMSchemaService> service = domMountPoint.getService(DOMSchemaService.class);
-            final Optional<DOMRpcService> domRpcServiceOptional = domMountPoint.getService(DOMRpcService.class);
-            if (domRpcServiceOptional.isPresent()) {
-                return Optional.of(new NetconfNmdaBaseServiceImpl(nodeId, domRpcServiceOptional.get(),
-                        service.orElseThrow().getGlobalContext()));
-            }
-        }
-        return Optional.empty();
-    }
-
-    private Optional<DOMMountPoint> getNetconfDOMMountPoint(NodeId nodeId) {
-        final YangInstanceIdentifier yangInstanceIdentifier = NetconfUtils.createNetConfNodeMountPointYII(nodeId);
-        return this.domMountPointService.getMountPoint(yangInstanceIdentifier);
     }
 }
